// mono-repo/packages/gantry-app/src/app/cds/dark-theme-matches/common/models/dark-theme-commom-matches.model.ts
export class DarkThemeCommonMatches {
    title?: string | null | undefined;
    status?: string | null | undefined;
    homePrice?: string | null | undefined;
    homeSelectionTitle?: string | null | undefined;
    marketVersesName?: string | null | undefined;
    awayPrice?: string | null | undefined;
    awaySelectionTitle?: string | null | undefined;
    drawPrice?: string | null | undefined;
    drawTitle?: string | null | undefined;
    drawSuspended?: string | null | undefined | boolean;
    drawMarketAvailable?: boolean;
    trimTitleRequired?: boolean;
}